requirements.txt
fastapi>=0.68.0,<0.69.0
python-dotenv
PyMySQL==1.0.2
PyJWT==2.4.0
uvicorn==0.17.6